+++ tests/testthat/test-is_golem.R
@@ -1,0 +1,8 @@
+test_that("is_golem works", {
+  expect_true(
+    is_golem(pkg)
+  )
+  expect_false(
+    is_golem(tempdir())
+  )
+})