+++ tests/testthat/test-zzzzzzzzzz.R
@@ -1,5 +1,7 @@
 # # For setting back old usethis settings
 try({
+  unlink("README.Rmd", TRUE, TRUE)
+  unlink("README.md", TRUE, TRUE)
   unlink(pkg, TRUE, TRUE)
   options("usethis.quiet" = old_usethis.quiet)
 })