+++ NEWS.md
@@ -1,5 +1,41 @@
 > Notes: the # between parenthesis referes to the related issue on GitHub, and the @ refers to an external contributor solving this issue.
 
+# 0.4.2 to 0.5.0
+
+## New functions
+
+* `is_golem()` tries to guess if the current folder is a `{golem}`-based app (#836)
+
+* `use_readme_rmd()` adds a {golem} specific README.Rmd (@ilyaZar, #1011)
+
+## New features / user visible changes
+
++ `add_fct()` now adds the skeleton for a function (#1004, @ilyaZar)
+
++ The module skeleton now stick to tidyverse style (#1019, @ni2scmn)
+
++ Better comments to `fill_desc()` in `01_start.R`  (#1021, @ilyaZar)
+
++ `01_start.R` now has a call to `usethis::use_git_remote()` (#1015, @ilyaZar)
+
++ Test for utilsui and server now has full code coverate (#1020, @ilyaZar)
+
++ When setting a new name, golem now browses tests & vignettes (#805, @ilyaZar)
+
+## Bug fix
+
++ Docker commands now take the `-it` flag so it can be killed with `^C` (#, @ivokwee)
+
++ add_module() now behaves correctly when trying to use `mod_mod_XXX` (#997, @ilyaZar)
+
+## Internal changes
+
++ All functions that requires to get a path now rely on get_golem_wd() (#1016, @ilyaZar)
+
+# 0.4.1
+
+This is an intermediate release after CRAN feedback.
+
 # 0.4.0
 
 ## New functions
@@ -32,10 +68,7 @@
 
 ## Internal changes
 
-add_dockerfile_with_renv now works well with uppercase in package name
-
-add_dockerfile_with_renv now works well with uppercase in package name
-
++ `add_dockerfile_with_renv` now works well with uppercase in package name
 
 # golem 0.3.5
 
@@ -55,7 +88,7 @@ Update in the tests for CRAN (skip not installed + examples).
 
 - `add_dockerfile`, `add_dockerfile_shinyproxy()` and `add_dockerfile_heroku()` now recommend to switch to their `_with_renv_` counterpart
 
-# golem 0.3.2 (CRAN VERSION)
+# golem 0.3.2
 
 ### Soft deprecated
 
@@ -131,7 +164,7 @@ Update in the tests for CRAN (skip not installed + examples).
 
 - `{testthat}` and `{rlang}` are no longer hard dependencies (#742)
 
-# golem 0.3.1 (CRAN Version)
+# golem 0.3.1
 
 ## New functions